Refactor Delivery service
